+++ messages/APRep.go
@@ -37,7 +37,7 @@ type EncAPRepPart struct {
 	CTime          time.Time           `asn1:"generalized,explicit,tag:0"`
 	Cusec          int                 `asn1:"explicit,tag:1"`
 	Subkey         types.EncryptionKey `asn1:"optional,explicit,tag:2"`
-	SequenceNumber int                 `asn1:"optional,explicit,tag:3"`
+	SequenceNumber int64               `asn1:"optional,explicit,tag:3"`
 }
 
 // Unmarshal bytes b into the APRep struct.
