+++ pac/signature_data_test.go
@@ -38,7 +38,7 @@ func TestPAC_SignatureData_Unmarshal_KDC_Signature(t *testing.T) {
 	}
 	sig, _ := hex.DecodeString("340be28b48765d0519ee9346cf53d822")
 	zeroed, _ := hex.DecodeString("76ffffff00000000000000000000000000000000")
-	assert.Equal(t, uint32(chksumtype.KERB_CHECKSUM_HMAC_MD5_UNSIGNED), k.SignatureType, "Server signature type not as expected")
+	assert.Equal(t, chksumtype.KERB_CHECKSUM_HMAC_MD5_UNSIGNED, k.SignatureType, "Server signature type not as expected")
 	assert.Equal(t, sig, k.Signature, "Server signature not as expected")
 	assert.Equal(t, uint16(0), k.RODCIdentifier, "RODC Identifier not as expected")
 	assert.Equal(t, zeroed, bz, "Returned bytes with zeroed signature not as expected")
